requirements.txt
pandas
numpy
scikit-learn
joblib

fastapi
uvicorn
pydantic

pytest
httpx

apache-airflow-providers-docker
prometheus_client
mlflow
xgboost